Move core functions to more appropriate module
