+++ src/poliastro/core/thrust/change_argp.py
@@ -3,8 +3,7 @@ from numba import njit as jit
 from numpy import cross
 from numpy.linalg import norm
 
-from poliastro.core.elements import rv2coe
-from poliastro.core.util import circular_velocity
+from poliastro.core.elements import circular_velocity, rv2coe
 
 
 @jit
